[Api][Zone] Change DataPersister to StateProcessor
